Функция инициализации перенесена в метод init контроллера, прямое обращение к приватным методам заменено обращением к ним через геттеры в соответствии с принципами ООП, используются атрибуты data для реализации сортировки
